chore: Move DataHolder into its own file
